video_generator/requirements.txt
# C-Cube Tutorial Video Generator - Required Packages

# Core video processing
moviepy==1.0.3
opencv-python==4.8.1.78
pillow==10.0.1
numpy==1.24.3

# Text-to-speech and audio
gTTS==2.4.0
pyttsx3==2.90
pydub==0.25.1

# Screen capture and automation
pyautogui==0.9.54
selenium==4.15.2
mss==9.0.1

# Graphics and animations
matplotlib==3.7.2
imageio==2.31.1

# Voice synthesis (advanced options)
# elevenlabs==0.2.26  # Uncomment if you want AI voice
# edge-tts==6.1.9     # Uncomment for Microsoft Edge voices

# Utilities
requests==2.31.0
json5==0.9.14
